pyproject.toml
[project]
name = "travel-planner-with-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geopy>=2.4.1",
    "google-adk>=1.19.0",
    "python-dotenv>=1.2.1",
]
